Split block and stream files using a command flag
